pyproject.toml
[tool.poetry]
name = "first-airflow-steps"
version = "0.1.0"
description = ""
authors = ["Rodrigo Coelho <rodrigo.coelho@hotmail.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
apache-airflow = "^2.7.3"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
